Change todo structure ( todo -completed )
