--- src/components/Text.tsx
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-
-class Text extends Component {
-  render() {
-    return <Text style={{ fontFamily: 'Noto' }}>{Children}</Text>
-  }
-}
-
-export default Text
