+++ src/pages/Todo.tsx
@@ -1,12 +1,8 @@
 import React from 'react'
 import {
-  Text,
   View,
   ListView,
   Image,
-  StyleSheet,
-  TouchableOpacity,
-  Dimensions,
 } from 'react-native'
 import { connect } from 'react-redux'
 import { Action, Todo } from '../reducers'
@@ -14,12 +10,11 @@ import PageLayout from '../components/PageLayout'
 import Header from '../components/Header'
 import FilterAndSearch from '../components/FilterAndSearch'
 import AddBlueberryBtn from '../components/AddBlueberryBtn'
-import TodoListItem from '../components/TodoListItem'
-import TodoListSectionHeader from '../components/TodoListSectionHeader'
-import { SwipeListView } from 'react-native-swipe-list-view'
+import TodoList from '../components/TodoList'
+import CompletedList from '../components/CompletedList'
 import AddTodoModal from '../components/AddTodoModal'
 import _ from 'lodash'
-import { mainColor } from '../config'
+
 
 namespace TodoComponent {
   export interface Props {
@@ -30,13 +25,14 @@ namespace TodoComponent {
     ds: any
     dataSource: any
     isAddMode: boolean
+    isTodoList: boolean
   }
 }
 
 class TodoComponent extends React.Component<
   TodoComponent.Props,
   TodoComponent.State
-> {
+  > {
   constructor(props) {
     super(props)
     const ds = new ListView.DataSource({
@@ -50,6 +46,7 @@ class TodoComponent extends React.Component<
         _.groupBy(this.props.todos, 'createdAt'),
       ),
       isAddMode: false,
+      isTodoList: true,
     }
   }
 
@@ -70,11 +67,11 @@ class TodoComponent extends React.Component<
             style={{ height: 23, width: 23, marginTop: 7 }}
           />
         ) : (
-          <Image
-            source={require('../assets/Global/todo_default.png')}
-            style={{ height: 17, width: 22, marginTop: 5 }}
-          />
-        )}
+            <Image
+              source={require('../assets/Global/todo_default.png')}
+              style={{ height: 17, width: 22, marginTop: 5 }}
+            />
+          )}
       </View>
     ),
   }
@@ -97,59 +94,23 @@ class TodoComponent extends React.Component<
           close={() => this.setState({ isAddMode: false })}
         />
         <Header />
-        <FilterAndSearch />
-        <AddBlueberryBtn onPress={() => this.setState({ isAddMode: true })} />
-        {this.props.todos.length === 0 ? (
-          <View style={styles.emptyBox}>
-            <View>
-              <Text style={styles.emptyText}>블루베리로</Text>
-              <Text style={styles.emptyText}>할 일을</Text>
-              <Text style={styles.emptyText}>시작해볼까요?</Text>
-            </View>
-            <View style={{ alignItems: 'flex-end' }}>
-              <TouchableOpacity
-                activeOpacity={0.8}
-                onPress={() => this.setState({ isAddMode: true })}
-              >
-                <Image source={require('../assets/Todo/blueberry_empty.png')} />
-              </TouchableOpacity>
-            </View>
+        <FilterAndSearch
+          changeTodoList={() => this.setState({ isTodoList: true })}
+          changeCompletedList={() => this.setState({ isTodoList: false })}
+          isTodoList={this.state.isTodoList}
+        />
+        {this.state.isTodoList ? (
+          <View>
+            <AddBlueberryBtn onPress={() => this.setState({ isAddMode: true })} />
+            <TodoList
+              dataSource={this.state.dataSource}
+              onPress={() => this.setState({ isAddMode: true })}
+            />
           </View>
         ) : (
-          <SwipeListView
-            dataSource={this.state.dataSource}
-            renderRow={todo => (
-              <TodoListItem
-                title={todo.title}
-                sessionCount={todo.sessionCount}
-              />
-            )}
-            disableRightSwipe
-            renderSectionHeader={(_, category) => (
-              <TodoListSectionHeader date={category} />
-            )}
-            renderHiddenRow={() => (
-              <View
-                style={{
-                  alignItems: 'center',
-                  flex: 1,
-                  flexDirection: 'row',
-                  justifyContent: 'flex-end',
-                  paddingLeft: 15,
-                  height: 50,
-                }}
-              >
-                <View style={{ backgroundColor: 'red', flex: 1, height: 50 }}>
-                  <Text>Delete</Text>
-                </View>
-                <View style={{ backgroundColor: 'blue' }}>
-                  <Text>Start</Text>
-                </View>
-              </View>
-            )}
-            rightOpenValue={-75}
-          />
-        )}
+            <CompletedList dataSource={this.state.dataSource} />
+          )}
+
       </PageLayout>
     )
   }
@@ -162,17 +123,4 @@ export default connect(
   undefined,
 )(TodoComponent)
 
-const styles = StyleSheet.create({
-  emptyText: {
-    // fontFamily: fontFamily.thin,
-    fontSize: 40,
-    color: mainColor.light,
-  },
-  emptyBox: {
-    justifyContent: 'space-between',
-    alignContent: 'space-between',
-    paddingHorizontal: 42,
-    paddingVertical: 70,
-    height: Dimensions.get('window').height - 240,
-  },
-})
+
